update RDM responder example
