+++ examples/idf_rdm_responder/CMakeLists.txt
@@ -1,4 +1,4 @@
 idf_component_register(
-    SRCS "idf_rdm_read.c"
+    SRCS "idf_rdm_responder.c"
     INCLUDE_DIRS ""
 )
